+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Update TypeScript type definitions.
 
 ## [1.3.3] - 2022-01-15