+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Update TypeScript type definitions.
+- Update type information in the documentation.
 
 ## [1.3.3] - 2022-01-15