+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Versioning].
 
 - Add escaping for Unix interpolation characters to `escape`/`escapeAll`.
 - Add escaping for Zsh wildcard characters to `escape`/`escapeAll`.
+- Update TypeScript type definitions.
+- Update type information in the documentation.
 
 ## [1.3.3] - 2022-01-15